Refatorações, Travis para branch dev, novas instruções pra Rodar, Nix movido para um diretório próprio
