+++ bot/__init__.py
@@ -1,23 +1,20 @@
 import asyncio
 import logging
 import sys
-from itertools import repeat
-from operator import eq
 
 from telethon import TelegramClient, errors, events
 from telethon.sessions import StringSession
-from telethon.tl.functions.messages import GetStickerSetRequest
-from telethon.tl.types import InputStickerSetID
 
 from .configuration import settings
-from .core import commands, default
+from .plugins.default import default
+from .plugins.user import commands
 
-logging_level = logging.INFO if settings.debug else logging.WARNING
+LOGGING_LEVEL = logging.INFO if settings.debug else logging.WARNING
 
 logging.basicConfig(
     stream=sys.stdout,
     format="[%(levelname) 5s/%(asctime)s] %(name)s: %(message)s",
-    level=logging_level,
+    level=LOGGING_LEVEL,
 )
 
 async_logger = logging.getLogger("asyncio")
@@ -25,8 +22,10 @@ async_logger.setLevel(logging.ERROR)
 
 SESSION = settings.session
 
-bot = TelegramClient(StringSession(SESSION), settings.api_id, settings.api_hash)
+Bot = TelegramClient(StringSession(SESSION), settings.api_id, settings.api_hash)
+
+SPAMMERS = {}
 
 # Setando os handlers do bot
-default(bot)
-commands(bot)
+default(Bot)
+commands(Bot)
