+++ bot/configuration/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseConfig
+
+settings = BaseConfig()
